Added more missing code
Game will not run, but I followed the book all the way and have all the scripts, prefabs and all the settings in the inspector according to the book. At this point, I give up on trying to figure out why my game wont run. The problem seemingly started from not being able to parse both the LayoutXML and CardXML